requirements.txt
fastapi
uvicorn[standard]
httpx
pytrends
sqlalchemy
asyncpg
pydantic
python-dotenv
google-ads
fastapi
uvicorn[standard]
httpx
pydantic
SQLAlchemy>=1.4
asyncpg
alembic
python-dotenv
pytrends
tqdm
apscheduler
pytz
psycopg2-binary
